[FEATURE] Perform Data Analytics on Output Generated JSON #8
Analysing JSON data from 2 different ROM images in the form of Pie chart and bar graph view of various parameters.

This implementation allows us to see comparative analysis of 2 binary files. With the help of pie chart and bar graphs, we can analyze various components of section types and file types.

Implementation for #8

Signed-off-by: Kumar, Shubham <shubham.k@intel.com>